+++ backend/src/baserow/contrib/database/table/cache.py
@@ -10,7 +10,7 @@ starts at 0 and is incremented every time:
 2. A process generates a new field_attrs for caching by querying the database.
 
 We then store cache field_attrs in the cache key:
-    `full_table_model_{table_id}_{model_version}_{BASEROW_VERSION}`
+    `full_table_model_{table_id}_{min_model_version}_{BASEROW_VERSION}`
 
 When we construct a model we:
 1. Get the latest model version
@@ -22,7 +22,7 @@ be sure concurrent changes to the model aren't going to overwrite each others
 changes to the cached field_attrs.
 """
 
-from typing import Dict, Any
+from typing import Dict, Any, Optional
 
 from django.conf import settings
 from django.core.cache import caches
@@ -37,50 +37,60 @@ def table_model_cache_version_key(table_id: int) -> str:
     return f"full_table_model_version_{table_id}_{BASEROW_VERSION}"
 
 
-def table_model_cache_entry_key(table_id: int, model_version: int) -> str:
-    return f"full_table_model_{table_id}_{model_version}_{BASEROW_VERSION}"
+def table_model_cache_entry_key(table_id: int) -> str:
+    return f"full_table_model_{table_id}_{BASEROW_VERSION}"
 
 
-def all_model_version_table_model_cache_entry_key(table_id: int) -> str:
-    return f"full_table_model_{table_id}_*"
-
-
-def get_latest_cached_model_field_attrs(table_id: int) -> Dict[str, Any]:
+def get_cached_model_field_attrs(
+    table_id: int, min_model_version: int
+) -> Optional[Dict[str, Any]]:
     """
+    :param min_model_version: The model version to lookup, if an older or no version is
+        found then None will be returned. Use get_current_cached_model_version to get
+        this value.
     :param table_id: The table to lookup any cached mode field attrs for.
     :return: The latest cached field attrs for the table's model or None if nothing has
-        been cached yet.
+        been cached yet or the attrs in the cache are older than the provided min
+        version.
+    """
+
+    cache_key = table_model_cache_entry_key(table_id)
+
+    cache_entry = generated_models_cache.get(cache_key)
+    if cache_entry and cache_entry["version"] >= min_model_version:
+        return cache_entry["field_attrs"]
+    else:
+        return None
+
+
+def get_current_cached_model_version(table_id: int) -> int:
+    """
+    Returns the current cached model version for the table id. This can be then used
+    when getting the cached model attrs using the get_cached_model_field_attrs method.
     """
 
     model_version_key = table_model_cache_version_key(table_id)
-    latest_model_version = generated_models_cache.get_or_set(
-        model_version_key, 0, timeout=None
-    )
-
-    cache_key = table_model_cache_entry_key(table_id, latest_model_version)
-
-    return generated_models_cache.get(cache_key)
+    return generated_models_cache.get_or_set(model_version_key, 0, timeout=None)
 
 
-def set_cached_model_field_attrs(table_id: int, field_attrs: Dict[str, Any]):
+def set_cached_model_field_attrs(
+    table_id: int, model_version: int, field_attrs: Dict[str, Any]
+):
     """
     Will increment the latest model version for table_id and store field_attrs in the
     cache entry for that model version.
 
     :param table_id: The table to lookup any cached mode field attrs for.
+    :param model_version: The version of the model being set in the cache.
     :param field_attrs: The field_attrs for table_id to cache.
     """
 
-    model_version_key = table_model_cache_version_key(table_id)
-    # We are setting a new version of the tables field_attrs, we incr here again to
-    # ensure `next_model_version` is unique to this process and no other concurrent
-    # process will ever override what we are about to set.
-    # If we did not do this two different processes could get the same
-    # `next_model_version` and then race to set field_attrs. Both processes could have
-    # queried the database at different times and constructed different field_attrs.
-    next_model_version = generated_models_cache.incr(model_version_key)
-    cache_key = table_model_cache_entry_key(table_id, next_model_version)
-    generated_models_cache.set(cache_key, field_attrs, timeout=None)
+    cache_key = table_model_cache_entry_key(table_id)
+    generated_models_cache.set(
+        cache_key,
+        {"field_attrs": field_attrs, "version": model_version},
+        timeout=None,
+    )
 
 
 def clear_generated_model_cache():
@@ -110,36 +120,62 @@ def invalidate_table_in_model_cache(
     other tables if need be.
     """
 
-    model_version_key = table_model_cache_version_key(table_id)
-    model_version = generated_models_cache.get_or_set(model_version_key, 0)
-
     if invalidate_related_tables:
-        _invalidate_all_related_models(table_id, model_version)
+        _invalidate_all_related_models(table_id)
 
-    if hasattr(generated_models_cache, "delete_pattern"):
-        all_versions_cache_key = all_model_version_table_model_cache_entry_key(table_id)
-        # By deleting all versions we clean up any old versions left in the cache.
-        generated_models_cache.delete_pattern(all_versions_cache_key)
-    else:
-        generated_models_cache.delete(
-            table_model_cache_entry_key(table_id, model_version)
-        )
-
-    generated_models_cache.incr(model_version_key)
-
-    return model_version
+    model_version_key = table_model_cache_version_key(table_id)
+    _incr_or_create_key_atomically_if_possible(model_version_key)
 
 
-def _invalidate_all_related_models(table_id: int, related_model_version: int):
+def _incr_or_create_key_atomically_if_possible(key_to_incr: str) -> int:
+    """
+    Attempts to increment the provided key if present or otherwise create it in an
+    atomic fashion.
+
+    If the cache backend does not support the "ignore_key_check" parameter required
+    to do an atomic incr/create then we fall back to a non-atomic incr and create.
+    """
+
+    try:
+        return generated_models_cache.incr(key_to_incr, ignore_key_check=True)
+    except TypeError:
+        # We are using a cache backend that doesn't support ignore_key_check so we can't
+        # do an atomic increment or create. Instead we do our best and incr/set
+        # separately
+        try:
+            return generated_models_cache.incr(key_to_incr)
+        except ValueError:
+            generated_models_cache.set(key_to_incr, 1, timeout=None)
+            return 1
+
+
+def _invalidate_all_related_models(table_id: int):
+    """
+    Given a table id looks up the latest cached field_attrs for it and loops over
+    all of the fields found in the field_attrs calling their
+    before_table_model_invalidated hook. Then in the link row field this hook will
+    recursively invalidate the linked table.
+
+    We need to do this as a change in one tables model might invalidate the cached
+    models of any related tables.
+
+    :param table_id: The table to invalidate all other table models which are related
+        via link row fields to this table via.
+    """
+
     from baserow.contrib.database.fields.registries import field_type_registry
 
-    this_version_cache_key = table_model_cache_entry_key(
-        table_id, related_model_version
-    )
-    field_attrs = generated_models_cache.get(this_version_cache_key)
-    if field_attrs is not None:
-        for field_obj in field_attrs.get("_field_objects", []).values():
-            field = field_obj.get("field", None)
-            if field is not None:
-                field_type = field_type_registry.get_by_model(field)
-                field_type.before_table_model_invalidated(field)
+    this_version_cache_key = table_model_cache_entry_key(table_id)
+
+    try:
+        field_objects = generated_models_cache.get(this_version_cache_key, {})[
+            "field_attrs"
+        ]["_field_objects"]
+    except KeyError:
+        field_objects = {}
+
+    for field_obj in field_objects.values():
+        field = field_obj.get("field", None)
+        if field is not None:
+            field_type = field_type_registry.get_by_model(field)
+            field_type.before_table_model_invalidated(field)